+++ rust/candid/src/pretty/candid.rs
@@ -71,15 +71,15 @@ fn ident_string(id: &str) -> String {
     }
 }
 
-pub fn pp_text(id: &str) -> RcDoc {
+pub fn pp_text(id: &str) -> RcDoc<'_> {
     RcDoc::text(ident_string(id))
 }
 
-pub fn pp_ty(ty: &Type) -> RcDoc {
+pub fn pp_ty(ty: &Type) -> RcDoc<'_> {
     pp_ty_inner(ty.as_ref())
 }
 
-pub fn pp_ty_inner(ty: &TypeInner) -> RcDoc {
+pub fn pp_ty_inner(ty: &TypeInner) -> RcDoc<'_> {
     use TypeInner::*;
     match ty {
         Null => str("null"),
@@ -107,8 +107,8 @@ pub fn pp_ty_inner(ty: &TypeInner) -> RcDoc {
         Record(ref fs) => {
             let t = Type(ty.clone().into());
             if t.is_tuple() {
-                let tuple = concat(fs.iter().map(|f| pp_ty(&f.ty)), ";");
-                kwd("record").append(enclose_space("{", tuple, "}"))
+                let fs = fs.iter().map(|f| pp_ty(&f.ty));
+                kwd("record").append(sep_enclose_space(fs, ";", "{", "}"))
             } else {
                 kwd("record").append(pp_fields(fs, false))
             }
@@ -130,18 +130,18 @@ pub fn pp_docs<'a>(docs: &'a [String]) -> RcDoc<'a> {
 /// This function is kept for backward compatibility.
 ///
 /// It is recommended to use [`pp_label_raw`] instead, which accepts a [`Label`].
-pub fn pp_label(id: &SharedLabel) -> RcDoc {
+pub fn pp_label(id: &SharedLabel) -> RcDoc<'_> {
     pp_label_raw(id.as_ref())
 }
 
-pub fn pp_label_raw(id: &Label) -> RcDoc {
+pub fn pp_label_raw(id: &Label) -> RcDoc<'_> {
     match id {
         Label::Named(id) => pp_text(id),
         Label::Id(_) | Label::Unnamed(_) => RcDoc::as_string(id),
     }
 }
 
-pub(crate) fn pp_field(field: &Field, is_variant: bool) -> RcDoc {
+pub(crate) fn pp_field(field: &Field, is_variant: bool) -> RcDoc<'_> {
     let ty_doc = if is_variant && *field.ty == TypeInner::Null {
         RcDoc::nil()
     } else {
@@ -150,12 +150,11 @@ pub(crate) fn pp_field(field: &Field, is_variant: bool) -> RcDoc {
     pp_label_raw(&field.id).append(ty_doc)
 }
 
-fn pp_fields(fs: &[Field], is_variant: bool) -> RcDoc {
-    let fields = fs.iter().map(|f| pp_field(f, is_variant));
-    enclose_space("{", concat(fields, ";"), "}")
+fn pp_fields(fs: &[Field], is_variant: bool) -> RcDoc<'_> {
+    sep_enclose_space(fs.iter().map(|f| pp_field(f, is_variant)), ";", "{", "}")
 }
 
-pub fn pp_function(func: &Function) -> RcDoc {
+pub fn pp_function(func: &Function) -> RcDoc<'_> {
     let args = pp_named_args(&func.args);
     let rets = pp_rets(&func.rets);
     let modes = pp_modes(&func.modes);
@@ -168,7 +167,7 @@ pub fn pp_function(func: &Function) -> RcDoc {
 /// Pretty-prints named arguments in the form of `(name1 : type1, name2 : type2)`.
 ///
 /// To print unnamed arguments, use [`pp_args`] instead.
-pub fn pp_named_args(args: &[ArgType]) -> RcDoc {
+pub fn pp_named_args(args: &[ArgType]) -> RcDoc<'_> {
     let args = args.iter().map(|arg| {
         if let Some(name) = &arg.name {
             pp_text(name).append(kwd(" :")).append(pp_ty(&arg.typ))
@@ -176,54 +175,49 @@ pub fn pp_named_args(args: &[ArgType]) -> RcDoc {
             pp_ty(&arg.typ)
         }
     });
-    let doc = concat(args, ",");
-    enclose("(", doc, ")")
+    sep_enclose(args, ",", "(", ")")
 }
 
 /// Pretty-prints arguments in the form of `(type1, type2)`.
 ///
 /// To print named arguments, use [`pp_named_args`] instead.
-pub fn pp_args(args: &[Type]) -> RcDoc {
-    let doc = concat(args.iter().map(pp_ty), ",");
-    enclose("(", doc, ")")
+pub fn pp_args(args: &[Type]) -> RcDoc<'_> {
+    sep_enclose(args.iter().map(pp_ty), ",", "(", ")")
 }
 
 /// Pretty-prints return types in the form of `(type1, type2)`.
-pub fn pp_rets(args: &[Type]) -> RcDoc {
+pub fn pp_rets(args: &[Type]) -> RcDoc<'_> {
     pp_args(args)
 }
 
-pub fn pp_mode(mode: &FuncMode) -> RcDoc {
+pub fn pp_mode(mode: &FuncMode) -> RcDoc<'_> {
     match mode {
         FuncMode::Oneway => RcDoc::text("oneway"),
         FuncMode::Query => RcDoc::text("query"),
         FuncMode::CompositeQuery => RcDoc::text("composite_query"),
     }
 }
-pub fn pp_modes(modes: &[FuncMode]) -> RcDoc {
+pub fn pp_modes(modes: &[FuncMode]) -> RcDoc<'_> {
     RcDoc::concat(modes.iter().map(|m| RcDoc::space().append(pp_mode(m))))
 }
 
 fn pp_service<'a>(serv: &'a [(String, Type)], docs: Option<&'a DocComments>) -> RcDoc<'a> {
-    let doc = concat(
-        serv.iter().map(|(id, func)| {
-            let doc = docs
-                .and_then(|docs| docs.lookup_service_method(id))
-                .map(|docs| pp_docs(docs))
-                .unwrap_or(RcDoc::nil());
-            let func_doc = match func.as_ref() {
-                TypeInner::Func(ref f) => pp_function(f),
-                TypeInner::Var(_) => pp_ty(func),
-                _ => unreachable!(),
-            };
-            doc.append(pp_text(id)).append(kwd(" :")).append(func_doc)
-        }),
-        ";",
-    );
-    enclose_space("{", doc, "}")
+    let methods = serv.iter().map(|(id, func)| {
+        let doc = docs
+            .and_then(|docs| docs.lookup_service_method(id))
+            .map(|docs| pp_docs(docs))
+            .unwrap_or(RcDoc::nil());
+        let func_doc = match func.as_ref() {
+            TypeInner::Func(ref f) => pp_function(f),
+            TypeInner::Var(_) => pp_ty(func),
+            _ => unreachable!(),
+        };
+        doc.append(pp_text(id)).append(kwd(" :")).append(func_doc)
+    });
+    sep_enclose_space(methods, ";", "{", "}")
 }
 
-fn pp_defs(env: &TypeEnv) -> RcDoc {
+fn pp_defs(env: &TypeEnv) -> RcDoc<'_> {
     lines(env.to_sorted_iter().map(|(id, ty)| {
         kwd("type")
             .append(ident(id.as_str()))
@@ -523,7 +517,7 @@ pub mod value {
         }
     }
 
-    fn pp_field(depth: usize, field: &IDLField, is_variant: bool) -> RcDoc {
+    fn pp_field(depth: usize, field: &IDLField, is_variant: bool) -> RcDoc<'_> {
         let val_doc = if is_variant && field.val == IDLValue::Null {
             RcDoc::nil()
         } else {
@@ -532,9 +526,9 @@ pub mod value {
         pp_label_raw(&field.id).append(val_doc)
     }
 
-    fn pp_fields(depth: usize, fields: &[IDLField]) -> RcDoc {
-        let fs = concat(fields.iter().map(|f| pp_field(depth, f, false)), ";");
-        enclose_space("{", fs, "}")
+    fn pp_fields(depth: usize, fields: &[IDLField]) -> RcDoc<'_> {
+        let fs = fields.iter().map(|f| pp_field(depth, f, false));
+        sep_enclose_space(fs, ";", "{", "}")
     }
 
     pub fn pp_char(v: u8) -> String {
@@ -545,7 +539,7 @@ pub mod value {
             format!("\\{v:02x}")
         }
     }
-    pub fn pp_value(depth: usize, v: &IDLValue) -> RcDoc {
+    pub fn pp_value(depth: usize, v: &IDLValue) -> RcDoc<'_> {
         use IDLValue::*;
         if depth == 0 {
             return RcDoc::as_string(format!("{v:?}"));
@@ -561,13 +555,13 @@ pub mod value {
                     RcDoc::as_string(format!("{v:?}"))
                 } else {
                     let values = vs.iter().map(|v| pp_value(depth - 1, v));
-                    kwd("vec").append(enclose_space("{", concat(values, ";"), "}"))
+                    kwd("vec").append(sep_enclose_space(values, ";", "{", "}"))
                 }
             }
             Record(fields) => {
                 if is_tuple(v) {
                     let fields = fields.iter().map(|f| pp_value(depth - 1, &f.val));
-                    kwd("record").append(enclose_space("{", concat(fields, ";"), "}"))
+                    kwd("record").append(sep_enclose_space(fields, ";", "{", "}"))
                 } else {
                     kwd("record").append(pp_fields(depth, fields))
                 }
@@ -579,11 +573,11 @@ pub mod value {
         }
     }
 
-    pub fn pp_args(args: &IDLArgs) -> RcDoc {
+    pub fn pp_args(args: &IDLArgs) -> RcDoc<'_> {
         let args = args
             .args
             .iter()
             .map(|v| pp_value(MAX_ELEMENTS_FOR_PRETTY_PRINT, v));
-        enclose("(", concat(args, ","), ")")
+        sep_enclose(args, ",", "(", ")")
     }
 }