+++ rust/candid_parser/tests/assets/ok/empty.d.ts
@@ -1,6 +1,6 @@
-import type { Principal } from '@dfinity/principal';
 import type { ActorMethod } from '@dfinity/agent';
 import type { IDL } from '@dfinity/candid';
+import type { Principal } from '@dfinity/principal';
 
 export type T = [T];
 export interface _SERVICE {
@@ -8,5 +8,7 @@ export interface _SERVICE {
   'g' : ActorMethod<[T], { 'a' : T }>,
   'h' : ActorMethod<[[T, never]], { 'a' : T } | { 'b' : {} }>,
 }
+export declare const idlService: IDL.ServiceClass;
+export declare const idlInitArgs: IDL.Type[];
 export declare const idlFactory: IDL.InterfaceFactory;
 export declare const init: (args: { IDL: typeof IDL }) => IDL.Type[];