+++ rust/candid_parser/tests/assets/ok/http_streaming.js
@@ -1,3 +1,54 @@
+import { IDL } from '@dfinity/candid';
+
+export const StreamingToken = IDL.Record({
+  'resource' : IDL.Text,
+  'index' : IDL.Nat,
+});
+export const StreamingCallbackHttpResponse = IDL.Record({
+  'token' : IDL.Opt(StreamingToken),
+  'body' : IDL.Vec(IDL.Nat8),
+});
+export const HeaderField = IDL.Tuple(IDL.Text, IDL.Text);
+export const HttpRequest = IDL.Record({
+  'url' : IDL.Text,
+  'method' : IDL.Text,
+  'body' : IDL.Vec(IDL.Nat8),
+  'headers' : IDL.Vec(HeaderField),
+});
+export const StreamingCallback = IDL.Func(
+    [StreamingToken],
+    [StreamingCallbackHttpResponse],
+    ['query'],
+  );
+export const StreamingStrategy = IDL.Variant({
+  'Callback' : IDL.Record({
+    'token' : StreamingToken,
+    'callback' : StreamingCallback,
+  }),
+});
+export const HttpResponse = IDL.Record({
+  'body' : IDL.Vec(IDL.Nat8),
+  'headers' : IDL.Vec(HeaderField),
+  'streaming_strategy' : IDL.Opt(StreamingStrategy),
+  'status_code' : IDL.Nat16,
+});
+
+export const idlService = IDL.Service({
+  'httpStreamingCallback' : IDL.Func(
+      [StreamingToken],
+      [StreamingCallbackHttpResponse],
+      ['query'],
+    ),
+  'http_request' : IDL.Func([HttpRequest], [HttpResponse], ['query']),
+  'upload' : IDL.Func(
+      [IDL.Text, IDL.Text, IDL.Vec(IDL.Nat8), IDL.Bool],
+      [],
+      [],
+    ),
+});
+
+export const idlInitArgs = [];
+
 export const idlFactory = ({ IDL }) => {
   const StreamingToken = IDL.Record({
     'resource' : IDL.Text,
@@ -31,6 +82,7 @@ export const idlFactory = ({ IDL }) => {
     'streaming_strategy' : IDL.Opt(StreamingStrategy),
     'status_code' : IDL.Nat16,
   });
+  
   return IDL.Service({
     'httpStreamingCallback' : IDL.Func(
         [StreamingToken],
@@ -45,4 +97,5 @@ export const idlFactory = ({ IDL }) => {
       ),
   });
 };
+
 export const init = ({ IDL }) => { return []; };