+++ rust/candid_parser/tests/assets/ok/inline_methods.d.ts
@@ -1,6 +1,6 @@
-import type { Principal } from '@dfinity/principal';
 import type { ActorMethod } from '@dfinity/agent';
 import type { IDL } from '@dfinity/candid';
+import type { Principal } from '@dfinity/principal';
 
 export type Fn = ActorMethod<[bigint], bigint>;
 export type Gn = Fn;
@@ -23,5 +23,7 @@ export interface _SERVICE {
   'high_order_fn_via_record' : ActorMethod<[R], bigint>,
   'high_order_fn_via_record_inline' : ActorMethod<[RInline], bigint>,
 }
+export declare const idlService: IDL.ServiceClass;
+export declare const idlInitArgs: IDL.Type[];
 export declare const idlFactory: IDL.InterfaceFactory;
 export declare const init: (args: { IDL: typeof IDL }) => IDL.Type[];