+++ rust/candid_parser/tests/assets/ok/management.d.ts
@@ -1,6 +1,6 @@
-import type { Principal } from '@dfinity/principal';
 import type { ActorMethod } from '@dfinity/agent';
 import type { IDL } from '@dfinity/candid';
+import type { Principal } from '@dfinity/principal';
 
 export type bitcoin_address = string;
 export type bitcoin_network = { 'mainnet' : null } |
@@ -171,5 +171,7 @@ export interface _SERVICE {
     undefined
   >,
 }
+export declare const idlService: IDL.ServiceClass;
+export declare const idlInitArgs: IDL.Type[];
 export declare const idlFactory: IDL.InterfaceFactory;
 export declare const init: (args: { IDL: typeof IDL }) => IDL.Type[];